requirements-dev.txt
transifex-client==0.14.2
sphinx
pep8
flake8
django-debug-toolbar==3.2
django-extensions
jupyter
black==20.8b1
